+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Pipeline step script generation uses ERB templates (@mkasztelnik)
 - Blood flow computation is self-contained - Ansys files are downloaded from git
   repository (@mkasztelnik)
+- Gitlab integration services (list branches/tags and download specific file) (@amber7b)
 
 ### Changed
 - Upgrade to ruby 2.4.1 (@mkasztelnik)